refactor: server.ts モジュール分割 (#11)
872行のモノリス server.ts を8ファイルに分割。
純粋なリファクタリングで機能変更なし。

- types.ts: Session, MessageRow 型定義
- tmux.ts: tmux操作（liveness cache, exec, notify）
- db.ts: DB初期化、共有状態アクセサ、クリーンアップ
- tools/register.ts: wire_register, wire_unregister
- tools/messaging.ts: wire_send, wire_receive, wire_broadcast, wire_ack
- tools/sessions.ts: wire_sessions, wire_status
- tools/control.ts: wire_thread, wire_control
- server.ts: エントリポイント（28行）
